Aloitettu kirjasto

+++ student/02/lotto/main.cpp
@@ -5,12 +5,12 @@ using namespace std;
 int main()
 {
 
-    long int pallojen_maara;
-    long int nostetut_pallot;
-    long int erotusten_faktori (int n, int p);
-    long int maaran_faktori (int n);
-    long int nostettujen_faktori (int n);
-    long int todennakoisyys (int n, int p);
+    long long int pallojen_maara;
+    long long int nostetut_pallot;
+    long long int erotusten_faktori (int n, int p);
+    long long int maaran_faktori (int n);
+    long long int nostettujen_faktori (int n);
+    long long int todennakoisyys (int n, int p);
 
     cout<<"Enter the total number of lottery balls: ";
     cin >>pallojen_maara;
@@ -30,9 +30,9 @@ int main()
 
 
 }
-long int maaran_faktori (int n)
+long long int maaran_faktori (int n)
 {
-   long int luku = 1;
+   long long int luku = 1;
    for (int i = 1; i < n;++i)
        {
        luku *= i+1;
@@ -40,9 +40,9 @@ long int maaran_faktori (int n)
 return luku;
 }
 
-long int nostettujen_faktori (int n)
+long long int nostettujen_faktori (int n)
 {
-   long int faktori= 1;
+   long long int faktori= 1;
    for (int i = 1; i < n;++i)
        {
        faktori *= i+1;
@@ -51,10 +51,10 @@ return faktori;
 
 }
 
-long int erotusten_faktori (int n, int p)
+long long int erotusten_faktori (int n, int p)
 {
-    long  int erotus = n-p;
-    long int faktori = 1;
+    long long int erotus = n-p;
+    long long int faktori = 1;
     for (int i = 1; i < erotus;++i)
         {
         faktori *= i+1;
@@ -62,13 +62,13 @@ long int erotusten_faktori (int n, int p)
     return faktori;
 
 }
- long int todennakoisyys (int n, int p)
+ long long int todennakoisyys (int n, int p)
 
 {
-  long int ylakerta = maaran_faktori(n);
-  long int erotus_faktori = erotusten_faktori(n, p);
-  long int nostettu_faktori = nostettujen_faktori(p);
-  long int tulos = ylakerta/(erotus_faktori * nostettu_faktori);
+  long long int ylakerta = maaran_faktori(n);
+  long long int erotus_faktori = erotusten_faktori(n, p);
+  long long int nostettu_faktori = nostettujen_faktori(p);
+  long long int tulos = ylakerta/(erotus_faktori * nostettu_faktori);
    return tulos;
 }
